8 utilize anyhow for better error messages (#9)
* feat: use `anyhow` for better Errors

As detailed in the rust CLI book, the crate `anyhow` allows for better
and more understandable error propagation, specifically due to its
ability to provide both context and the original error.

* style: format code using `cargo fmt`

* feat: improve error context using `anyhow`.

The new implementations of the `anyhow` package use it to better extent
in order for the context to be referencing the errors thrown by the
different parts of the project. Giving the user a better insight into
the nature of the error.

---------

Co-authored-by: BassedWarrior <bassedwarrior724@outlook.com>