3 create tests (#17)
* refactor: move CLI parsing to own module file.

* refactor: modularize new project creation

Move new project creation logic to its own file and module in order to have
more testable code which is split up amongst various different functions
that better delimit the logic of each of `cppargo`'s subcommands.

* refactor: modularize project building

Move building logic to its own module file for better organization,
splitting of functions and more unit-testable code.

* refactor: modularize project running

Move running logic to its own module file for better organization,
splitting of functions and more unit-testable code.

* test: unit testing for new project creation

Refactor some functions to be able to implement unit testing for the
creation of a new project. Both dealing with the creation of directories
and files, as well as the contents of the main "Hello World!" `main.cpp`
program that the new project creates.

* test: implement unit testing for project building

Implement unit tests for the building of a `cppargo` project. This goes
over the discovery of the file structure, as well as the creation of the
intended binaries with the proper name.

* test: implement unit testing for project running

Ensure the correct finding of the project binary to be run by `cppargo`.

* fix: fix spelling mistakes

* style: improve print messages

Clarify the fact that not finding the `src` and `target` directories
implies that `cppargo` is not bein run at the root of a `cppargo`
project.

* fix: add newline at end of sample hello world file

* test: cli testing for end-user behaviour

Test correct behaviour of all 3 subcommands specific to `cppargo` from
the end-user perspective by running the compiled binary utilizing the
`assert_cmd`.