22 handle directories in source file discovery (#25)
* feat: handle src subdirectories for `.cpp` files

- Utilize HashSets instead of Vecs for the found src files found, avoiding
    any sort of order considerations and automatically avoiding any sort of
    repeated entries.
- Recursively scour through the different subdirectories which might be
    found inside the src directory for finding `.cpp` files.

* feat: improve find_src_files unit test

Extend the unit test for finding the src `.cpp` files in the given src
directory by including all assortment of distinct locations where the
`.cpp` files are expected to be found. This should also include the
traversal of subdirectories and the ignoring of empty subdirectories.

* fix: use hashset src_files for unit test

Due to the new data type for src_files handling, a fix was required to
no longer use Vec.

* refactor: utilize constants for file content

Ensure file contents are identical by utilizing constants to avoid any
sort of discrepancy between the different tests.

* refactor: apply clippy::pedantic lints