feat: create `Cppargo.toml` manifest at project root
Create a basic `Cppargo.toml` file to mark the project root such that it
will be possible to utilize `cppargo` from inside a project, and not
rely on the command being called when at the root of the project. Which
currently depends on a `src` and `target` directories being present.
Both ambiguous and not that robust. With this change, it will be a lot
more permissive to use `cppargo` no matter where in the project file
structure it is run from.