feat: ensure target directory exists
Due to the new implementation of the `new` subcommand for `cppargo`, the
`target` directory is not created when creating a new project, hence
verification that this directory is present is needed.

In case the directory already exists, be it by an already run `build` or
`run` subcommand, or by manual creation by the user, it does nothing and
proceeds. And in the case that it doesn't, it attempts to create it.