test: improve finding project root testing
Due to the new implementation of the way the root of the project is
determined and found, it is crucial to adjust the testing of such
functionality in the conttext of adding the `Cppargo.toml` manifest.
Which in this case is further separated into distinct cases within a
sub-module of the testing module for better organization of the
distinct tests.