test: correct target dir testing
The new implementation no longer relies on the target dir being created
by the `new` subcommand. Which in the case of testing the `build`
subcommand also implies it should not be present in test cases which
assume no previous `build` subcommands have been run. Which is also the
reason to include the new test, which expects the target directory and
a binary to exist already.