7 usage from inside project subdirectories (#26)
* feat: create `Cppargo.toml` manifest at project root

Create a basic `Cppargo.toml` file to mark the project root such that it
will be possible to utilize `cppargo` from inside a project, and not
rely on the command being called when at the root of the project. Which
currently depends on a `src` and `target` directories being present.
Both ambiguous and not that robust. With this change, it will be a lot
more permissive to use `cppargo` no matter where in the project file
structure it is run from.

* refactor: stop creating empty `target` subdirectory

Rather than creating an empty `target` subdirectory from the project
root with the sole purpose of signaling the root of a project, the use
of the new `Cppargo.toml` renders the need for such a directory
obsolete, and should therefore only be created when building the binary
from the project.

* feat: utilize manifest for project root and name

Search for the project root recursively up to filesystem root looking
for a `Cppargo.toml` file. From which it is able to gather the project's
root directory, as well as the project's name. Which is important for
building the binary file in the `target` subdirectory.

* feat: ensure target directory exists

Due to the new implementation of the `new` subcommand for `cppargo`, the
`target` directory is not created when creating a new project, hence
verification that this directory is present is needed.

In case the directory already exists, be it by an already run `build` or
`run` subcommand, or by manual creation by the user, it does nothing and
proceeds. And in the case that it doesn't, it attempts to create it.

* refactor: move manifest parsing to own function

For testing purposes, it is benefitial to have this be a standalone
function that can be tested individually without any other
considerations.

* refactor: simplify building function

Since the target directory path and the project name are both known
beforehand, and they are only used to build the output_path of the built
binary, then it's much more concise to have them unified in a single
parameter which is dealt with individually in the funciton logic.

* refactor: follow cargo-clippy pedantic suggestions

Implement suggestions made by cargo-clippy with the pedantic preset for
linting set. This makes the code more idiomatic to Rust.

* refactor: improve error messages

Due to the change in the way `cppargo` determines where the root of a
project is, instead of only specifying that it's not being run from the
root of a project, it helps to better understand that it is now reliant
on a `Cppargo.toml` file being present in order to gather the root
directory of the project.

* test: ensure tests comply with implementation

In the case of the renaming of the function to `main`, do so. And in the
case of the assertion for the contents of the file, utilize better
syntax for the content of the file, not only its existance.

* test: improve finding project root testing

Due to the new implementation of the way the root of the project is
determined and found, it is crucial to adjust the testing of such
functionality in the conttext of adding the `Cppargo.toml` manifest.
Which in this case is further separated into distinct cases within a
sub-module of the testing module for better organization of the
distinct tests.

* test: modularize finding src files testing cases

Instead of having a single big unit test which tests all possible
different cases, build a sub-module for testing that function and have
separate unit tests for each case that should be tested.

* refactor: add context to failed parsing of manifest

* test: add testing for project name parsing

Include two cases for the parsing of the name of the project from the
manifest. Testing both the success case and the missing name case, which
would entail a failure to parse the manifest in the first place.

* test: add testing to ensuring target dir exists

Test both cases of the target directory already existing, as well as the
case for the target dir needing to be created.

* test: update build_src_files test

Improve the clarity of the test with better variables, as well as adjust
the call to the `build_src_files()` function, which changed to be made
more simple.

* test: update main unit test

Include the necessity for a manifest being present in order to find the
root of the project, and properly build the project, as well as change
the name of the function following the adjustment suggested by clippy.

* test: rename project_dir to project_root

Since the new implementation relies more on the existance of a root for
the project, even though it is still a directory, I believe it is
clearer to have this be the name of the variable.

* test: include existance of project manifest

Include the creation of a project manifest to be used in the tests,
which is now mandatory in order to have the tests run properly.

* refactor: improve error message clarity

It is no longer required that the commands be run from the project root,
hence it is more clear to specify that the current directory is not
inside a `cppargo` project, which is determined by the failure to find
a project manifest `Cppargo.toml` file.

* test: update tests to match new error message

* test: correct target dir testing

The new implementation no longer relies on the target dir being created
by the `new` subcommand. Which in the case of testing the `build`
subcommand also implies it should not be present in test cases which
assume no previous `build` subcommands have been run. Which is also the
reason to include the new test, which expects the target directory and
a binary to exist already.