docs: update `README.md` with new functionality
Update the _usage_ section of the `README.md` for the new functionality
explanation of what each different subcommand does with the new changes.
This includes a more in-depth, maybe wiki-style redaction and could be
helped by introducing a separate Wiki, and a Quickstart section for the
`README.md` to have a more concise way to understand how to use
`cppargo`.