7 usage from inside project subdirectories (#27)
* refactor: rename `project_dir` to `project_root`

In accordance to the convention started in the rest of the files which
more closely relates to the new functionality of finding the project
root by looking for the `Cppargo.toml`.

* docs: adapt and improve the `--help` menu strings.

Due to the changes in implementation and behaviour, the `--help`
required some adjustments to better explain and reflect the new
behaviour of the different commands.

* docs: update `README.md` with new functionality

Update the _usage_ section of the `README.md` for the new functionality
explanation of what each different subcommand does with the new changes.
This includes a more in-depth, maybe wiki-style redaction and could be
helped by introducing a separate Wiki, and a Quickstart section for the
`README.md` to have a more concise way to understand how to use
`cppargo`.

* docs: update `README.md` with Quickstart section

As mentioned in the previous commit (84378a2), the new redaction for the
usage section of the `README.md` reads more like a Wiki than a quick
summary, so a Quickstart section could help impatient readers just get
going.