refactor: modularize ensuring successful project creation
In order to better integrate different functionalities that should be
able to be tested independently, and only added on when the entire
functionality is to be tested, the function for ensuring the proper
project creation is split. Further modules for VCS or other extra
functionality can then be appended to the comprehensive function.
Without the need to rewrite the code for all the tests which are
intended to test the entire available functionality.