21 implement vcs with git (#28)
* style: cargo fmt format change

* feat: initialize `git` repo with new projects

If `git` is found, then use it to initialize the created project
directory as a `git` reposotory as a Version Control System (VCS).

* test: unit test for proper VCS initialization.

This test only succeeds iff `git` is found and the repo is properly
initialized. Else it fails.

* refactor: modularize ensuring successful project creation

In order to better integrate different functionalities that should be
able to be tested independently, and only added on when the entire
functionality is to be tested, the function for ensuring the proper
project creation is split. Further modules for VCS or other extra
functionality can then be appended to the comprehensive function.
Without the need to rewrite the code for all the tests which are
intended to test the entire available functionality.

* test: test VCS interaction with the whole system.

Include it as part of the functionality to be tested when testing
normally, and also include a separate test for its own case where the
program should have no problems if the repo can't be initialized because
`git` is not found.

* docs: provide documentation and help messages

Improve the documentation to reflect the new behaviour by both modifying
the `README.md` file as well as the help message output when using the
`--help` flag or the `help` subcommand.