refactor(build): better errors for name parsing
While parsing the manifest in order to find the name, a lot of different
things can go wrong, or be wrong. The old implementation assumed that
nothing, except for the lack of a name, could be wrong. This new
implementation fixes that and clarifies what the mistake in the manifest
syntax is.