build: update all dependencies to latest versions (#30)
Assuming that the integration of the latest versions would mean that the
various dependencies could only receive bug fixes, enhancements, and
security vulnerabilities addressed, I'm trusting that updating
everything is just good for the project. Although it may not be
necessary at all. And so long as nothing breaks, I'll take it as a win.