test(build): unit tests for manifest parsing
Due to the refactoring done to the way the project name is gathered.
Splitting it into two different functions, each in charge of a different
part of parsing the manifest to obtain the project name, the unit tests
used for such functionality were expanded and improved.

Now it can check for more errors, and give better test coverage to the
logic of both functions.