test(build): correct integration test
According to the new behaviour, rather than looking for the error to be
that no `.cpp` files were found, simply look for the error to be that
there is no `src/main.cpp` file present. Which is what the new
implementation does. Which also ensures that at least a single `.cpp`
file is present.