24 require srcmaincpp for building (#32)
* feat(build): fail if `main.cpp` is missing.

Before even attempting to look for any `.cpp` file, ensure that the
`src/main.cpp` file is present. This with the aim of enforcing standards
and conventions when it comes to file structure.

* test(build): correct integration test

According to the new behaviour, rather than looking for the error to be
that no `.cpp` files were found, simply look for the error to be that
there is no `src/main.cpp` file present. Which is what the new
implementation does. Which also ensures that at least a single `.cpp`
file is present.

* docs(readme): specify importance of `src/main.cpp`

The requirement that a `src/main.cpp` file be present in order for
`cppargo` to even attempt to look for any other `.cpp` files, and
subsequently call `g++` to compile the project was added to the
`README.md` file in order to document the enforcement of this file
structure standard, product of this latest refactoring.